Logic class, append_bband(), and fixes
Added the logic class which will be used for algo logic. In this I made the append_bband() function that will calculate and return the bollinger bands from a dataset within a certain window. Also fixed problems with the standard_dev() and moving_average() functions that caused them to incorrectly inflate the moving average/standard deviation